+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/to_struct.rs
@@ -3,11 +3,8 @@ use crate::conversions::v3_0::to_rust_type::{
     object_to_field_shapes, DefinitionShape, FieldShape, TypeShape,
 };
 use crate::conversions::Result;
-use crate::targets::rust_type::{DataType, StructDef, StructField, StructFieldName};
-use openapi_types::v3_0::{
-    ReferenceObject, RequiredSchemaFields, SchemaCase, SchemaFieldName, SchemaObject,
-    SchemaProperties,
-};
+use crate::targets::rust_type::{DataType, StructDef};
+use openapi_types::v3_0::{ReferenceObject, SchemaCase, SchemaFieldName, SchemaObject};
 use SchemaCase::{Reference, Schema};
 
 pub(super) fn to_struct(name: SchemaFieldName, object: SchemaObject) -> Result<DefinitionShape> {
@@ -34,45 +31,6 @@ pub(super) fn to_struct(name: SchemaFieldName, object: SchemaObject) -> Result<D
         DefinitionShape::Fixed(def.into())
     };
     Ok(shape)
-}
-
-/// SchemaProperties -> Vec<StructField>
-pub(super) struct FieldsFactory {
-    pub required: Option<RequiredSchemaFields>,
-}
-
-impl FieldsFactory {
-    pub fn apply(self, props: SchemaProperties) -> Result<Vec<FieldShape>> {
-        props
-            .into_iter()
-            .map(|(name, case)| self.to_field(name, case))
-            .collect()
-    }
-
-    fn to_field(&self, name: SchemaFieldName, case: SchemaCase) -> Result<FieldShape> {
-        match shape_type(case)? {
-            TypeShape::Fixed(mut data_type) => {
-                if !self.is_required(&name) {
-                    data_type = DataType::Option(Box::new(data_type));
-                }
-                Ok(FieldShape::Fixed(StructField {
-                    name: StructFieldName::new(name),
-                    data_type,
-                }))
-            }
-            type_shape => Ok(FieldShape::InProcess {
-                name: StructFieldName::new(name),
-                type_shape,
-            }),
-        }
-    }
-
-    fn is_required(&self, name: &SchemaFieldName) -> bool {
-        match &self.required {
-            Some(required) => required.contains(name.as_ref()),
-            None => false,
-        }
-    }
 }
 
 pub(super) fn shape_type(schema_case: SchemaCase) -> Result<TypeShape> {
